+++ postfix/src/global/header_token.h
@@ -24,14 +24,14 @@ typedef struct HEADER_TOKEN {
     int     type;			/* see below */
     union {
 	const char *value;		/* just a pointer, not a copy */
-	int     offset;			/* index into token buffer */
+	ssize_t offset;			/* index into token buffer */
     }       u;				/* indent beats any alternative */
 } HEADER_TOKEN;
 
 #define HEADER_TOK_TOKEN	256
 #define HEADER_TOK_QSTRING	257
 
-extern int header_token(HEADER_TOKEN *, int, VSTRING *, const char **, const char *, int);
+extern ssize_t header_token(HEADER_TOKEN *, ssize_t, VSTRING *, const char **, const char *, int);
 
 /* LICENSE
 /* .ad
