+++ postfix/src/global/mail_trigger.c
@@ -10,7 +10,7 @@
 /*	const char *class;
 /*	const char *service;
 /*	const char *request;
-/*	int	length;
+/*	ssize_t	length;
 /* DESCRIPTION
 /*	mail_trigger() wakes up the specified mail subsystem, by
 /*	sending it the specified request.
@@ -67,7 +67,7 @@
 /* mail_trigger - trigger a service */
 
 int     mail_trigger(const char *class, const char *service,
-		             const char *req_buf, int req_len)
+		             const char *req_buf, ssize_t req_len)
 {
     struct stat st;
     char   *path;
