+++ postfix/src/global/smtp_stream.h
@@ -33,9 +33,9 @@ extern void smtp_timeout_setup(VSTREAM *, int);
 extern void PRINTFLIKE(2, 3) smtp_printf(VSTREAM *, const char *,...);
 extern void smtp_flush(VSTREAM *);
 extern int smtp_fgetc(VSTREAM *);
-extern int smtp_get(VSTRING *, VSTREAM *, int);
-extern void smtp_fputs(const char *, int len, VSTREAM *);
-extern void smtp_fwrite(const char *, int len, VSTREAM *);
+extern int smtp_get(VSTRING *, VSTREAM *, ssize_t);
+extern void smtp_fputs(const char *, ssize_t len, VSTREAM *);
+extern void smtp_fwrite(const char *, ssize_t len, VSTREAM *);
 extern void smtp_fputc(int, VSTREAM *);
 
 extern void smtp_vprintf(VSTREAM *, const char *, va_list);
