+++ postfix/src/local/resolve.c
@@ -88,7 +88,7 @@ int     deliver_resolve_tree(LOCAL_STATE state, USER_ATTR usr_attr, TOK822 *addr
     char   *myname = "deliver_resolve_tree";
     RESOLVE_REPLY reply;
     int     status;
-    int     ext_len;
+    ssize_t ext_len;
     char   *ratsign;
 
     /*
